+++ lib/app/budgets/components/budget_evolution_chart.dart
@@ -5,7 +5,7 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:monekin/app/stats/utils/common_axis_titles.dart';
-import 'package:monekin/app/stats/widgets/fund_evolution_line_chart.dart';
+import 'package:monekin/app/stats/widgets/fund_evolution_info.dart';
 import 'package:monekin/core/models/budget/budget.dart';
 import 'package:monekin/i18n/generated/translations.g.dart';
