+++ lib/app/stats/widgets/fund_evolution_info.dart
